Updated the main program to use the new interfaces as a test harness. Updated a few variables as the correct endpoint was not passed correctly to interface functions. Made it so that buffer sizes are ints and not longs any more, cant see that operations in the foreseeable feature will need to go over 4GB.

+++ Source/SharpBattleNet.Framework.Networking/Connection/IConnection.cs
@@ -57,7 +57,7 @@ namespace SharpBattleNet.Framework.Networking.Connection
         /// The remote address to send the buffer to if this is a connectionless
         /// protocol.
         /// </param>
-        void Send(byte[] buffer, long bufferLenght = 0, EndPoint destination = null);
+        void Send(byte[] buffer, int bufferLenght = 0, EndPoint destination = null);
 
         void StartReceiving();
     }
